refactor(base): aligned base class usage and minor code cleansing
