+++ src/base/component.ts
@@ -1,16 +1,17 @@
-import { TypeScriptProject } from 'projen/lib/typescript';
+import { TypeScriptProjectBase } from './project';
 
 /**
  * Base class for managing project component configurations.
+ * @abstract
  */
 export abstract class Component {
-  protected project: TypeScriptProject;
+  protected project: TypeScriptProjectBase;
 
   /**
    * Initializes the component for a specified project.
    * @param project The project to configure the component for.
    */
-  constructor(project: TypeScriptProject) {
+  constructor(project: TypeScriptProjectBase) {
     this.project = project;
   }
 
@@ -18,31 +19,36 @@ export abstract class Component {
    * Adds template files to the component's project directory.
    * @protected
    */
-  protected addTemplates(): void {}
+  // eslint-disable-next-line prettier/prettier
+  protected addTemplates(): void { }
 
   /**
    * Configures settings specific to the component within the project.
    * @protected
    */
-  protected addSettings(): void {}
+  // eslint-disable-next-line prettier/prettier
+  protected addSettings(): void { }
 
   /**
    * Adds npm scripts specific to the component within the project.
    * @protected
    */
-  protected addScripts(): void {}
+  // eslint-disable-next-line prettier/prettier
+  protected addScripts(): void { }
 
   /**
    * Installs development dependencies for the component in the project.
    * @protected
    */
-  protected addDevDependencies(): void {}
+  // eslint-disable-next-line prettier/prettier
+  protected addDevDependencies(): void { }
 
   /**
    * Adds custom entries to the `.gitattributes` file for the component in the project.
    * @protected
    */
-  protected addGitAttributes(): void {}
+  // eslint-disable-next-line prettier/prettier
+  protected addGitAttributes(): void { }
 
   /**
    * Initialize component configuration for the project.
