+++ src/base/npm.ts
@@ -1,16 +1,17 @@
-import { TypeScriptProject } from 'projen/lib/typescript';
 import { Component } from './component';
 import { ProjenStandardScript } from '../types';
+import { TypeScriptProjectBase } from './project';
 
 /**
  * Base class for NPM Package component implementing all relevant configuration.
+ * @abstract
  */
 export abstract class NpmPackageBase extends Component {
   /**
    * Initializes the base NPM Package component.
    * @param project The project to configure NPM for.
    */
-  constructor(project: TypeScriptProject) {
+  constructor(project: TypeScriptProjectBase) {
     super(project);
   }
