+++ src/github-action/project.ts
@@ -21,7 +21,5 @@ export class GitHubActionProject extends TypeScriptProjectBase {
     new Husky(this);
     new CommitLint(this);
     new SampleCode(this);
-
-    // this.postSynthesize(() => console.log('hello'));
   }
 }
