+++ src/jsii/commitlint.ts
@@ -1,5 +1,4 @@
-import { TypeScriptProject } from 'projen/lib/typescript';
-import { CommitLintBase } from '../base/commitlint';
+import { CommitLintBase, TypeScriptProjectBase } from '../base';
 import { LintStagedConfig } from '../types';
 
 /**
@@ -12,7 +11,7 @@ export class CommitLintJsii extends CommitLintBase {
    * to ensure that all necessary configuration steps are applied.
    * @param project The project to configure CommitLint for.
    */
-  constructor(project: TypeScriptProject) {
+  constructor(project: TypeScriptProjectBase) {
     super(project);
     this.initialize();
   }
@@ -33,5 +32,6 @@ export class CommitLintJsii extends CommitLintBase {
    * @protected
    * @override
    */
-  protected addGitAttributes(): void {}
+  // eslint-disable-next-line prettier/prettier
+  protected addGitAttributes(): void { }
 }
