+++ src/jsii/vscode.ts
@@ -1,5 +1,4 @@
-import { TypeScriptProject } from 'projen/lib/typescript';
-import { VSCodeBase } from '../base';
+import { VSCodeBase, TypeScriptProjectBase } from '../base';
 
 /**
  * VSCode component implementing all relevant configuration for the Jsii project.
@@ -11,7 +10,7 @@ export class VSCodeJsii extends VSCodeBase {
    * to ensure that all necessary configuration steps are applied.
    * @param project The project to configure VSCode Package for.
    */
-  constructor(project: TypeScriptProject) {
+  constructor(project: TypeScriptProjectBase) {
     super(project);
     this.initialize();
   }
